Removing ComplexModelDTOModelBinder
